PLANET-5192 Ensure sed matches regardless of space
* Sed expected a space between the package name and the colon, however
this space is not there for some of the repos. The result was that sed
didn't replace anything, causing further jobs to not have the intended
version of the plugins. This fixes the issue for now, however depending
on regex substitution is error prone so we should do this in a more
robust way.