Fetch zip file for tags to use composer's artifact type
Since we already detect tags at this step, besides skipping building the
assets, we can also fetch the zip file from Github release already. We
can then use [composer's
artifact](https://getcomposer.org/doc/05-repositories.md#artifact)
repository type in the next step to just install that zip from the local
filesystem.

The only caveat of using artifacts is that `composer.json` file in that
zip needs to have the proper version. To achieve that, we inject the
version when we create the zip (see
[https7c790e110b471e510de07e1d7218b96055440407](https://github.com/greenpeace/planet4-plugin-gutenberg-blocks/commit/7c790e110b471e510de07e1d7218b96055440407).

We need also to make sure test instances work with this, where we
already use a different repository type (package). Since we use the
version "7" convention, this adds an extra check to skip building assets
but without trying to fectch a zip file; composer will take care of that
in the next step.