PLANET-6044 Pick available instance even when there is no ticket
Ref: https://jira.greenpeace.org/browse/PLANET-6044

---

This is needed to be able to use a test instance on PRs even when there
is no ticket (eg. forks)

To avoid a race condition that can result in using a test instance in
multiple PRs, instead of sorting, when there is no ticket we randomize
the list of instances to minimize chances of picking up the same instance.