Use DEPLOYED revisions also for rollback
Currently the rollback script only looks for `SUPERSEDED` revisions,
which in cases of failed deployments will skip the last successfully
deployed and go to the one before that.

Even in cases where the last revision is successfully deployed, but we
do want to rollback, this still works since we use:
`previous=${2:-$((current - 1))}` right before the look up loop.